backend/requirements.txt
# Python Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
pydantic[email]>=2.6.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Google OAuth & Gmail API (Phase 1.2)
google-auth>=2.25.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
requests>=2.31.0
httpx>=0.25.0

# For future units (Phase 1.3+)
# langchain
# langchain-google-genai
# langgraph

